added a bunch of wrapper functions for installers (mostly for atom)

+++ lib/cli_config/12-prezto.zsh
@@ -1,10 +1,17 @@
-setopt EXTENDED_GLOB
-for rcfile in "${ZDOTDIR:-$HOME}"/.zprezto/runcoms/^README.md(.N); do
-  ln -s "$rcfile" "${ZDOTDIR:-$HOME}/.${rcfile:t}"
-done
+#!/usr/bin/env zsh
 
+if [ ! -h $HOME/.zlogin]; then
+  setopt EXTENDED_GLOB
+  for rcfile in "${ZDOTDIR:-$HOME}"/.zprezto/runcoms/^README.md(.N); do
+    ln -s "$rcfile" "${ZDOTDIR:-$HOME}/.${rcfile:t}"
+  done
+fi
 # Install .zshrc
-rm -rf $HOME/.zshrc
-ln -s $(pwd)/.osx-prime/dotfiles/zshrc $HOME/.zshrc
-rm -rf $HOME/.zpreztorc
-ln -s $(pwd)/.osx-prime/dotfiles/zpreztorc $HOME/.zpreztorc
+if [ ! -h $HOME/.zshrc]; then
+  rm -rf $HOME/.zshrc
+  ln -s $(pwd)/.osx-prime/dotfiles/zshrc $HOME/.zshrc
+fi
+if [ ! -h $HOME/.zpreztorc ]; then
+  rm -rf $HOME/.zpreztorc
+  ln -s $(pwd)/.osx-prime/dotfiles/zpreztorc $HOME/.zpreztorc
+fi
